♻️ refactor/KIKI-31 : 폴더 재구조화
