--- src/main/java/com/jiyoung/kikihi/global/auth/jwt/user/JWTUserDetailsService.java
@@ -1,4 +1,0 @@
-package com.jiyoung.kikihi.global.auth.jwt.user;
-
-public class JWTUserDetailsService {
-}
