--- src/main/java/com/jiyoung/kikihi/global/auth/jwt/util/JWTExtractor.java
@@ -1,91 +1,0 @@
-package com.jiyoung.kikihi.global.auth.jwt.util;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.JwtParser;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.Date;
-import java.util.Optional;
-import java.util.UUID;
-
-import static com.jiyoung.kikihi.global.auth.jwt.util.JWTProvider.ACCESS_TOKEN_SUBJECT;
-
-@Component
-public class JWTExtractor {
-
-    private static final String BEARER = "Bearer ";
-    private static final String ID_CLAIM = "id";
-    private static final String EMAIL_CLAIM = "email";
-    private static final String ROLE_CLAIM = "role";
-
-    @Value("${kikihi.jwt.key}")
-    private String SECRET_KEY;
-
-    private Key key;
-
-    @PostConstruct
-    public void init() {
-        this.key = Keys.hmacShaKeyFor(SECRET_KEY.getBytes());
-    }
-
-    // 토큰 추출
-    public Optional<String> extractToken(HttpServletRequest request) {
-        return Optional.ofNullable(request.getHeader(ACCESS_TOKEN_SUBJECT))
-                .filter(token -> token.startsWith(BEARER))
-                .map(token -> token.replace(BEARER, ""));
-    }
-
-    // 사용자 정보 추출
-    public UUID getId(String token) {
-        return getIdFromToken(token, ID_CLAIM);
-    }
-
-    public String getEmail(String token) {
-        return getClaimFromToken(token, EMAIL_CLAIM);
-    }
-
-    public String getRole(String token) {
-        return getClaimFromToken(token, ROLE_CLAIM);
-    }
-
-    public Boolean isExpired(String token) {
-        Claims claims = parseClaims(token);
-        return claims.getExpiration().before(new Date());
-    }
-
-    private String getClaimFromToken(String token, String claimName) {
-        Claims claims = parseClaims(token);
-        return claims.get(claimName, String.class);
-    }
-
-    private UUID getIdFromToken(String token, String claimName) {
-        Claims claims = parseClaims(token);
-        return claims.get(claimName, UUID.class);
-    }
-
-    private Claims parseClaims(String token) {
-        JwtParser parser = Jwts.parserBuilder()
-                .setSigningKey(key)
-                .build();
-        return parser.parseClaimsJws(token).getBody();
-    }
-
-    // JWT 유효성 검증
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token);
-            return true;
-        } catch (Exception e) {
-            return false;
-        }
-    }
-}
