--- src/main/java/com/jiyoung/kikihi/global/auth/jwt/util/RedisUtil.java
@@ -1,50 +1,0 @@
-package com.jiyoung.kikihi.global.auth.jwt.util;
-
-import com.jiyoung.kikihi.global.common.response.CustomException;
-import com.jiyoung.kikihi.global.common.response.ErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Component;
-
-import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-
-@Component
-@RequiredArgsConstructor
-public class RedisUtil {
-
-    @Value("${kikihi.jwt.refresh.expiration}")
-    private Long refreshTokenExpiration;
-
-    private final static String TOKEN_FORMAT = "refreshToken:%s";
-    private final RedisTemplate<String, String> redisTemplate;
-
-
-    public void setRefreshToken(UUID id, String value) {
-        String key = String.format(TOKEN_FORMAT, id);
-        redisTemplate.opsForValue().set(key, value, refreshTokenExpiration, TimeUnit.MILLISECONDS);
-    }
-
-    public Object getRefreshToken(UUID id) {
-        String key = String.format(TOKEN_FORMAT, id);
-        Object getObjecet = redisTemplate.opsForValue().get(key);
-
-        if (getObjecet == null) {
-            throw new CustomException(ErrorCode.REFRESH_TOKEN_NOT_FOUND);
-        }
-
-        return getObjecet;
-    }
-
-    public boolean hasRefreshTokenKey(Long id) {
-        String key = String.format(TOKEN_FORMAT, id);
-        return Boolean.TRUE.equals(redisTemplate.hasKey(key));
-    }
-
-    public boolean deleteRefreshToken(Long id) {
-        String key = String.format(TOKEN_FORMAT, id);
-        return Boolean.TRUE.equals(redisTemplate.delete(key));
-    }
-
-}
