--- src/main/java/com/jiyoung/kikihi/global/common/response/ApiResponse.java
@@ -1,29 +1,0 @@
-package com.jiyoung.kikihi.global.common.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import io.micrometer.common.lang.Nullable;
-import org.springframework.http.HttpStatus;
-
-public record ApiResponse<T>(
-        @JsonIgnore
-        HttpStatus httpStatus,
-        boolean success,
-        @Nullable T data,
-        @Nullable ExceptionDto error
-) {
-
-    public static <T> ApiResponse<T> ok(@Nullable final T data) {
-        return new ApiResponse<>(HttpStatus.OK, true, data, null);
-    }
-
-    public static <T> ApiResponse<T> created(@Nullable final T data) {
-        return new ApiResponse<>(HttpStatus.CREATED, true, data, null);
-    }
-
-    public static <T> ApiResponse<T> fail(final CustomException e) {
-        return new ApiResponse<>(
-                e.getErrorCode().getHttpStatus(), false, null, ExceptionDto.of(e.getErrorCode(), e.getMessage()) // 예외 메시지 추가
-        );
-    }
-
-}
