--- src/main/java/com/jiyoung/kikihi/global/common/response/CustomException.java
@@ -1,19 +1,0 @@
-package com.jiyoung.kikihi.global.common.response;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public CustomException(ErrorCode errorCode, String message) {
-        super(message);
-        this.errorCode = errorCode;
-    }
-}
-
