+++ src/main/java/com/jiyoung/kikihi/global/response/page/PageRequest.java
@@ -1,0 +1,21 @@
+package com.jiyoung.kikihi.global.response.page;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import lombok.extern.slf4j.Slf4j;
+
+@Data
+@SuperBuilder
+@AllArgsConstructor
+@NoArgsConstructor
+@Slf4j
+public class PageRequest {
+    @Builder.Default
+    private int page = 1;
+
+    @Builder.Default
+    private int size = 10;
+}
