+++ src/main/java/com/jiyoung/kikihi/platform/application/in/auth/AuthUseCase.java
@@ -1,6 +1,6 @@
 package com.jiyoung.kikihi.platform.application.in.auth;
 
-import com.jiyoung.kikihi.global.auth.oauth2.kakao.KaKaoDto;
+import com.jiyoung.kikihi.security.oauth2.kakao.KaKaoDto;
 import com.jiyoung.kikihi.platform.domain.user.User;
 
 public interface AuthUseCase {
