+++ src/main/java/com/jiyoung/kikihi/security/jwt/dto/JWTTokenDto.java
@@ -1,0 +1,14 @@
+package com.jiyoung.kikihi.security.jwt.dto;
+
+
+import lombok.Builder;
+
+@Builder
+public record JWTTokenDto(
+        String accessToken,
+        String refreshToken
+) {
+    public static JWTTokenDto of(String accessToken, String refreshToken) {
+        return new JWTTokenDto(accessToken, refreshToken);
+    }
+}
