+++ src/main/java/com/jiyoung/kikihi/security/jwt/user/JWTUserDetailsService.java
@@ -1,0 +1,4 @@
+package com.jiyoung.kikihi.security.jwt.user;
+
+public class JWTUserDetailsService {
+}
