+++ src/main/java/com/jiyoung/kikihi/security/jwt/util/JWTProvider.java
@@ -1,0 +1,58 @@
+package com.jiyoung.kikihi.security.jwt.util;
+
+import com.jiyoung.kikihi.platform.domain.user.Role;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import java.util.Date;
+import java.util.UUID;
+
+@Component
+public class JWTProvider {
+
+    public static final String ACCESS_TOKEN_SUBJECT = "Authorization";
+    public static final String REFRESH_TOKEN_SUBJECT = "RefreshToken";
+    private static final String ID_CLAIM = "id";
+    private static final String EMAIL_CLAIM = "email";
+    private static final String ROLE_CLAIM = "role";
+    private static final String ROLE_PREFIX = "ROLE_";
+    private static final String DUMMY_EMAIL = "dummy_email";
+
+    @Value("${kikihi.jwt.key}")
+    private String SECRET_KEY;
+
+    @Value("${kikihi.jwt.access.expiration}")
+    private Long accessTokenExpiration;
+
+    @Value("${kikihi.jwt.refresh.expiration}")
+    private Long refreshTokenExpiration;
+
+    // Access Token 생성
+    public String generateAccessToken(UUID userId, String email, Role role) {
+        return Jwts.builder()
+                .claim(ID_CLAIM, userId)
+                .claim(EMAIL_CLAIM, email)
+                .claim(ROLE_CLAIM, ROLE_PREFIX + role)
+                .claim("tokenType", "ACCESS_TOKEN")
+                .setSubject(userId.toString())
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + accessTokenExpiration)) // 예: 15분
+                .signWith(SignatureAlgorithm.HS256, SECRET_KEY)
+                .compact();
+    }
+
+    // Refresh Token 생성
+    public String generateRefreshToken(UUID userId, String email, Role role) {
+        return Jwts.builder()
+                .claim(ID_CLAIM, userId)
+                .claim("tokenType", "REFRESH_TOKEN")
+                .setSubject(userId.toString())
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + refreshTokenExpiration)) // 예: 7일
+                .signWith(SignatureAlgorithm.HS256, SECRET_KEY)
+                .compact();
+    }
+
+
+}
